capsules/translator/requirements.txt
openai>=1.12.0
pyyaml>=6.0.1
jsonschema>=4.19.0
requests>=2.31.0
